something resembling a webcam config whosit
